feat: wire batch enrich to Gemini Batch API
