+++ src/brainlayer/vector_store.py
@@ -170,9 +170,11 @@ class VectorStore(SearchMixin, KGMixin, SessionMixin):
             ("tags", "TEXT"),
             ("tag_confidence", "REAL"),
             ("summary", "TEXT"),
+            ("summary_v2", "TEXT"),
             ("importance", "REAL"),
             ("intent", "TEXT"),
             ("enriched_at", "TEXT"),
+            ("enrichment_version", "TEXT DEFAULT '1.0'"),
             ("primary_symbols", "TEXT"),
             ("resolved_query", "TEXT"),
             ("key_facts", "TEXT"),
